requirements.txt
Django==4.2
djangorestframework==3.14
psycopg2-binary==2.9.6
celery==5.3.0
django-mptt==0.14.0
africastalking==1.2.5
gunicorn==20.1.0
drf-yasg==1.21.5
python-dotenv==1.0.0
pytest==7.4.0
pytest-django==4.5.2
django-oauth-toolkit==2.3.0
djangorestframework-simplejwt==5.3.0
python-jose==3.3.0
django-guardian==2.4.0